pyproject.toml
[tool.poetry]
name = "sensor_service"
version = "0.1.0"
description = ""
authors = ["Emmanuel Kinyanjui"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic = "^2.10.4"
google-cloud-pubsub = "^2.27.1"
python-dotenv = "^1.0.1"
huey = "^2.5.2"
redis = "^5.2.1"
pytest = "^8.3.4"
httpx = "^0.28.1"
sqlalchemy-utils = "^0.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
